fix feat: Fix "model already exists" bug and add block diagram screenshot feature
Fixes a bug where you couldn't ChatGPT couldn't create two Simulink models with the same name if the first model wasn't closed. Now any generated Simulink models are closed when running code from ChatGPT. Also, adds a feature that takes a screenshot of the simulink block diagram and saves it as an image. These images are displayed the same way as Figures in the Chat Window..